refactor(stellar): rename attest module to attestation for consistency
- Rename attest.rs to attestation.rs to align with domain terminology
- Update module imports to use attestation instead of attest
- Maintain all existing functionality and API surface
