+++ contracts/stellar/protocol/src/instructions/mod.rs
@@ -1,10 +1,10 @@
 pub mod schema;
-pub mod attest;
+pub mod attestation;
 pub mod delegation;
 pub mod crypto;
 
 pub use schema::{register_schema};
-pub use attest::{attest, get_attestation, revoke_attestation, list_attestations};
+pub use attestation::{attest, get_attestation, revoke_attestation, list_attestations};
 pub use delegation::{
     attest_by_delegation, 
     revoke_by_delegation
